corrected 1st task
