--- Practice/lecture14.Task1/Class2.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace lecture14.Task1
-{
-    public void SearchEqualRectangles()
-    {
-        
-        //for (int i = 0; i < a.RectangleArray().GetUpperbound
-    }
-}
